fix: all 553 tests pass — symlink config, cross-platform path, clippy fixes

+++ crates/swarm-engine/src/main.rs
@@ -360,7 +360,7 @@ async fn decision_cycle<P: alloy::providers::Provider>(
     };
 
     // Affective Intelligence — EWMA confidence & risk appetite from trade history
-    let ewma_conf = ewma_confidence(&pnl_outcomes.iter().map(|p| *p as f64).collect::<Vec<_>>(), 0.9);
+    let ewma_conf = ewma_confidence(&pnl_outcomes.iter().map(|p| *p).collect::<Vec<_>>(), 0.9);
     let risk_app = risk_appetite(affective.drawdown_state, 0.15); // 15% max DD
     tracing::info!("🧬 Affective: EWMA_conf={:.3} risk_appetite={:.3} streak={}",
         ewma_conf, risk_app, affective.consecutive_losses);
@@ -572,7 +572,7 @@ async fn decision_cycle<P: alloy::providers::Provider>(
         {
             let pe = paper.lock().unwrap();
             if pe.pnl_history.len() >= 5 {
-                let history: Vec<f64> = pe.pnl_history.iter().map(|p| *p as f64).collect();
+                let history: Vec<f64> = pe.pnl_history.iter().map(|p| *p).collect();
                 let latest_pnl = history.last().copied().unwrap_or(0.0);
                 let anomaly = hive_intel::anomaly::detect_anomaly(latest_pnl, &history);
                 if anomaly.severity != hive_intel::anomaly::AnomalySeverity::Normal {
@@ -791,7 +791,7 @@ async fn main() {
             if !pe.pnl_history.is_empty() {
                 let s = pe.stats();
                 t.paper_stats = Some(telemetry::PaperStats {
-                    total_trades: s.total_trades as u64,
+                    total_trades: s.total_trades,
                     win_rate: s.win_rate,
                     total_pnl: s.total_pnl,
                     max_drawdown: s.max_drawdown,
